fix: add correlationId to notification events
adds correlationId with the value of the task definition ID to all
notification events sent to notifo so that the correlationId can be used
to find notifications to check their status.

updates the notifo http client, because it was missing correlationId
fields previously.

also add a missing error check in the get scheduled notifications
function.